// React/src/models/User.ts
export interface User {
    id?: string;
    name?: string;
    username?: string;
    email?: string;
    password?: string;
    password2?: string;
    description?: string;
    location?: string;
    webpage?: string;
    picture?: string;
    backImage?: string;
    error?: string;
    followed?: number;
    followers?: number;
    tweets?: number
}